pyproject.toml
[project]
name = "sale-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "apache-airflow>=3.1.8",
    "apache-airflow-providers-google>=10.0.0",
    "google-cloud-storage",
    "google-cloud-bigquery",
    "google-cloud-dataproc",
    "cryptography",
]

[tool.uv]
package = false
override-dependencies = [
    "ray==2.44.1",
]
